fixed mme driver in x64 fixed DWORD_PTR problems in midi driver. removed some longs fixed a runtime check (from last commit of dsp) Proper readHeader for ITModule2

+++ trunk/psycle-core/src/psycle/core/song.cpp
@@ -309,7 +309,7 @@ bool CoreSong::IffAlloc(Instrument::id_type instrument,const char * str) {
 	return true;
 }
 
-bool CoreSong::WavAlloc(Instrument::id_type iInstr, bool bStereo, long iSamplesPerChan, const char * pathToWav) {
+bool CoreSong::WavAlloc(Instrument::id_type iInstr, bool bStereo, int32_t iSamplesPerChan, const char * pathToWav) {
 	assert(iSamplesPerChan < (1 << 30)); // Since in some places, signed values are used, we cannot use the whole range.
 	DeleteLayer(iInstr);
 	_pInstrument[iInstr]->waveDataL = new std::int16_t[iSamplesPerChan];
